Interpret scripts with desc=None as non-scripts
If desc=None in a Python file, then it will be marked as a non-script,
meaning is not listed as a script (not even with --list-all-scripts) and
is an error to try to call it as a script.

This is useful to add general Python modules acting as general libraries
and not intended to be used as scripts.