Don't ignore --oauth-token option
Most command-line arguments override configuration values by sharing the
same key internally and then updating the config dict globally.

Unfortunately the command-line variable for the OAuth token is named
`oauth_token` while the one in the config is named `oauthtoken`, so this
overriding mechanism fails for that particular value.

This commit just stores the CLI variable with the same name as the
config variable.

Fixes #2.