backup: Rewrite for org/user backup and better status
The whole status/progress management was re-done and the whole logic to
do the backup was rewritten to be able to backup full organizations and
users, recursively.

Now also the tool is verbose by default, and --verbose was replaced by
--quiet to make it silent.

Finally, no more per-repo zips are created, if no --file-name is given,
backup.zip will be used and all the backups are stored in one file.