Add request methods that also return the response(s)
This is also needed to properly handle ETags to cache requests. This way
users can also inspect response headers.

The new methods are: json_req_full(), get_full(), put_full(), etc. and they
return a tuple where the first element is a list of responses objects (more
than one can be returned if the request was paginated) and the second one
the parsed JSON object.